requirements.txt
google-generativeai
selenium
webdriver-manager
ffmpeg-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
google-cloud-storage
